requirements.txt
requests
beautifulsoup4
selenium
python-dotenv
pandas
openpyxl
matplotlib
seaborn